+++ web/admin/.storybook/main.ts
@@ -2,6 +2,7 @@ import type { StorybookConfig } from '@storybook/react-vite';
 
 const config: StorybookConfig = {
   stories: ['../src/**/*.stories.@(js|jsx|mjs|ts|tsx)'],
+  staticDirs: ['../public'],
   addons: ['@storybook/addon-essentials', '@storybook/addon-interactions', '@storybook/addon-links'],
   framework: { name: '@storybook/react-vite', options: {} },
 };